feat: Add user authentication

+++ app/Http/Controllers/ChecklistController.php
@@ -9,6 +9,10 @@ use Illuminate\Support\Facades\Auth;
 
 class ChecklistController extends Controller
 {
+    public function __construct() {
+        $this->middleware('auth', ['except' => ['index','show']]);
+    }
+
     public function index()
     {
         $checklists = Checklist::simplePaginate(10);
@@ -19,22 +23,27 @@ class ChecklistController extends Controller
     public function show($id)
     {
         $checklist = Checklist::findOrFail($id);
+        $id = Auth::id();
 
-        return view('checklist/show', compact('checklist'));
+
+        return view('checklist/show', compact('checklist', 'id'));
     }
 
     public function edit($id)
     {
         $checklist = Checklist::findOrFail($id);
 
-        $auth = Auth::user();
+        $this->authorize('edit', $checklist);
 
-        return view('checklist/edit', compact('checklist', 'auth'));
+        return view('checklist/edit', compact('checklist'));
     }
 
     public function update(Request $request, $id)
     {
         $checklist = Checklist::findOrFail($id);
+
+        $this->authorize('edit', $checklist);
+
         $checklist->checks = $request->checks;
         $checklist->description = $request->description;
         $checklist->name = $request->name;
@@ -46,6 +55,9 @@ class ChecklistController extends Controller
     public function destroy($id)
     {
         $checklist = Checklist::findOrFail($id);
+
+        $this->authorize('edit', $checklist);
+
         $checklist->delete();
 
         return redirect("/checklist");
